chore(cli): split 40_testing
