+++ cli/Cargo.toml
@@ -80,6 +80,7 @@ clap = { version = "=4.4.17", features = ["env", "string"] }
 clap_complete = "=4.4.7"
 clap_complete_fig = "=4.4.2"
 color-print = "0.3.5"
+constcat.workspace = true
 console_static_text.workspace = true
 dashmap = "5.5.3"
 data-encoding.workspace = true
